Shrink Timing struct from 16 to 8 bytes
Use u32 for delay_millis (supports ~49 days) and f32 for value,
halving per-record memory during long benchmark runs.

Made-with: Cursor

+++ src/common.rs
@@ -47,8 +47,8 @@ const BOOL_TRUE_RATIO: f64 = 0.7;
 
 #[derive(Debug, Clone)]
 pub struct Timing {
-    pub delay_millis: f64, // milliseconds
-    pub value: f64,
+    pub delay_millis: u32,
+    pub value: f32,
 }
 
 pub fn create_zipf(vocab_size: usize) -> rand_distr::Zipf<f64> {
